move all Chat components into a folder for easy management
