+++ src/App.js
@@ -11,48 +11,11 @@ import Feed from './FeedPage/FeedPage';
 import ProfilePage from './ProfilesPage/ProfilePage';
 import CreateEvents from './CreateEventsPage/CreateEvents';
 import EditPage from './EditPage/EditPage';
-import {Widget, addResponseMessage, handleNewUserMessage} from 'react-chat-widget';
-import logo from './images/Women.png';
-import Cable from 'actioncable';
-import LoginModal from './common/LoginModal';
+import ChatWidget from './ChatWidget/ChatWidget';
 
 class App extends Component {
 
-  //User types in new message in chat box
-  handleNewUserMessage = (newMessage) => {
-    // Now send the message to the backend API
-    this.chats.create(newMessage);
-    //Update the state of current message
-    this.state.currentMessage = newMessage;
-  }
-  componentDidMount() {
-    //Add welcome message at the beginning
-    addResponseMessage("Welcome to this awesome chat!");
-  }
-
-  getResponseMessage(message) {
-    //Make sure not to display own message in chat logs
-    if (this.state.currentMessage === message) {
-
-      //Reset the current message to null and return empty string
-      this.state.currentMessage = '';
-      return '';
-    } else {
-      //If it's not an own message, add to the chat log
-      addResponseMessage(message);
-
-      //Reset the current message to null
-      this.state.currentMessage = '';
-
-    }
-  }
-
   componentWillMount() {
-    //Create a Socket Connection for chat function
-    this.createSocket();
-    this.setState({currentMessage: ''});
-    this.setState({userLoggedIn: false});
-
     var config = {
       apiKey: "AIzaSyBErciM-iyLeO2x7c9Ly4G2JbQRbAadOnc",
       authDomain: "womenincomedy-cd5b5.firebaseapp.com",
@@ -62,38 +25,6 @@ class App extends Component {
       messagingSenderId: "1001680813005"
     };
     firebase.initializeApp(config);
-  }
-
-  //Prepare the Action Cable socket for chat function
-  createSocket() {
-    let cable = Cable.createConsumer('ws://localhost:9000/cable');
-
-    //Create chat function
-    this.chats = cable.subscriptions.create({
-      channel: 'ChatChannel'
-    }, {
-      connected: () => {},
-      received: (data) => {
-        this.getResponseMessage(data.content);
-      },
-      create: function(chatContent) {
-        this.perform('create', {content: chatContent});
-      }
-    });
-  }
-
-  renderChatWidget() {
-
-    //Obtain user id
-    const userId = sessionStorage.getItem('userId');
-
-    if (!userId) {
-      return <LoginModal/>
-    } else {
-      this.state.userLoggedIn = true;
-      return (<Widget handleNewUserMessage={this.handleNewUserMessage} profileAvatar={logo} title="Women In Comedy" subtitle="Chat Message"/>)
-
-    }
   }
 
   render() {
@@ -108,11 +39,9 @@ class App extends Component {
             <Route path="/home" component={HomePage}></Route>
             <Route path='/activities' component={ActivityPage}></Route>
             <Route path='/eventsfeed/:id' component={EventsFeed}></Route>
-            {/* <Route exact path='/profile/:id' component={ProfilePage}></Route>
-            <Route path='/profile/:id/edit' component={EditPage}></Route> */
-            }
+            <Route path='/profile/:id' component={ProfilePage}></Route>
           </Switch>
-          {this.renderChatWidget()}
+          <ChatWidget />
 
         </div>
       </MuiThemeProvider>
